requirements.txt
langchain
langgraph
openai
arxiv
faiss-cpu
sentence-transformers
pydantic
tqdm
pandas
scikit-learn
matplotlib
langchain-openai
langchain-huggingface
transformers
torch
arxiv
requests
pymupdf
chromadb
sentence-transformers
pydantic>=2.0
PyYAML
markdown2
weasyprint
biopython
networkx
python-slugify
pytest
ollama